requirements.txt
Appium-Python-Client>=5.2.4

# pytest + plugins
pytest==9.0.1
pytest-variables>=3.1.0
pytest-html>=4.1.1
pytest-metadata>=3.1.1
pytest-xdist>=3.5.0

# BrowserStack SDK
browserstack-sdk==1.31.0